requirements_pi.txt
# Raspberry Pi 4 inference dependencies
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
picamera2>=0.3.12
RPi.GPIO>=0.7.0
